requirements.txt
openai==0.27.6
pdfminer.six==20221105
python-dotenv==1.0.0
